organize promoted related utils
